refactor(arkitect): keep template around as a class field
Previously `makeRenderer` created a `Renderer` sub-class with a custom constructor passing forward the template received by `makeRenderer`. This kept the scope around and required an extra function to be constructed.

After this change, the sub-class is constructed without a custom constructor (therefore `Renderer`'s constructor is re-used directly). The template is now stored in a field on the new sub-class, and `Renderer` will access that field when copying the reference.